Major overhaul to use two threads
Use lazy_static for p5 API and thread_local for GL API. GL calls will
be executed from the main thread as GL windows have to be created on the
main thread on macOS. The main thread loops listening for closures being
sent over a channel. A second thread is spawned to run the main render
loop and code paths from here will send closures to trigger GL calls on
the main thread.